fix compillation waning (we use -Wall)

+++ src/algorithms/fardetectors/FarDetectorTransportationPostML.cc
@@ -80,7 +80,7 @@ void FarDetectorTransportationPostML::process(
   edm4eic::MutableReconstructedParticle particle;
 
   // Iterate over the prediction_tensor_data in steps of three
-  for (std::size_t i = 0; i < prediction_tensor.getShape(0); i++) {
+  for (std::size_t i = 0; i < static_cast<std::size_t>(prediction_tensor.getShape(0)); i++) {
 
     std::size_t base_index = i * 3;
